Add BigCalendar component for rendering calendar
