+++ package.json
@@ -8,6 +8,7 @@
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "dayjs": "^1.11.8",
     "react": "18.2.0",
+    "react-big-calendar": "^1.8.1",
     "react-dom": "18.2.0",
     "react-virtualized-auto-sizer": "^1.0.20",
     "tslib": "2.5.0"
@@ -27,6 +28,7 @@
     "@types/jest": "^29.5.2",
     "@types/lodash": "^4.14.195",
     "@types/node": "^18.16.16",
+    "@types/react-big-calendar": "^1.6.4",
     "@types/webpack-env": "^1.18.1",
     "@typescript-eslint/eslint-plugin": "^5.59.9",
     "copy-webpack-plugin": "^11.0.0",
