+++ provisioning/dashboards/weekly.json
@@ -68,11 +68,15 @@
       "options": {
         "annotations": true,
         "autoScroll": false,
+        "calendarType": "library",
         "colorField": "Color",
+        "colors": "frame",
         "descriptionField": "description",
         "displayTime": false,
         "endTimeField": "end",
-        "labelFields": ["labels"],
+        "labelFields": [
+          "labels"
+        ],
         "quickLinks": false,
         "textField": "text",
         "timeField": "start"
@@ -95,37 +99,73 @@
                 "config": {},
                 "name": "text",
                 "type": "string",
-                "values": ["Meeting", "Hard Work", "This", "Planning", "First in line"]
+                "values": [
+                  "Meeting",
+                  "Hard Work",
+                  "This",
+                  "Planning",
+                  "First in line"
+                ]
               },
               {
                 "config": {},
                 "name": "description",
                 "type": "string",
-                "values": ["First event", "Working hard", "work", "What to do?", "Event"]
+                "values": [
+                  "First event",
+                  "Working hard",
+                  "work",
+                  "What to do?",
+                  "Event"
+                ]
               },
               {
                 "config": {},
                 "name": "start",
                 "type": "time",
-                "values": [1670710154622, 1670609069806, 1670609069806, 1670809175025, 1660610154622]
+                "values": [
+                  1670710154622,
+                  1670609069806,
+                  1670609069806,
+                  1670809175025,
+                  1660610154622
+                ]
               },
               {
                 "config": {},
                 "name": "end",
                 "type": "time",
-                "values": [1670809017462, null, 1670609079806, null, null]
+                "values": [
+                  1670809017462,
+                  null,
+                  1670609079806,
+                  null,
+                  null
+                ]
               },
               {
                 "config": {},
                 "name": "labels",
                 "type": "string",
-                "values": ["hello", "hard,work", "this,that,who?", "", ""]
+                "values": [
+                  "hello",
+                  "hard,work",
+                  "this,that,who?",
+                  "",
+                  ""
+                ]
               },
               {
                 "config": {},
                 "name": "Color",
                 "type": "number",
-                "values": [30, 100, 30, 70, 60]
+                "values": [
+                  30,
+                  100,
+                  30,
+                  70,
+                  60
+                ]
               }
             ],
             "meta": {},
@@ -185,7 +225,8 @@
             "mode": "absolute",
             "steps": [
               {
-                "color": "green"
+                "color": "green",
+                "value": null
               },
               {
                 "color": "red",
@@ -252,6 +293,6 @@
   "timezone": "",
   "title": "Weekly",
   "uid": "pKxZUM5Vz",
-  "version": 3,
+  "version": 1,
   "weekStart": ""
 }
