+++ src/components/CustomCalendar/components/Day/Day.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import AutoSizer from 'react-virtualized-auto-sizer';
 import { cx } from '@emotion/css';
 import { useStyles2, useTheme2 } from '@grafana/ui';
-import { TestIds } from '../../constants';
-import { Styles } from '../../styles';
-import { CalendarEvent } from '../../types';
+import { TestIds } from '../../../../constants';
+import { Styles } from '../../../../styles';
+import { CalendarEvent } from '../../../../types';
 import { CalendarEntry } from '../CalendarEntry';
 
 /**
