+++ src/components/CustomCalendar/components/DayDrawer/DayDrawer.test.tsx
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { TestIds } from '../../constants';
-import { CalendarEvent } from '../../types';
+import { TestIds } from '../../../../constants';
+import { CalendarEvent } from '../../../../types';
 import { DayDrawer } from './DayDrawer';
 
 /**
