+++ src/components/LibCalendar/index.ts
@@ -1,0 +1,1 @@
+export * from './LibCalendar';
