+++ src/components/index.ts
@@ -1,5 +1,2 @@
-export * from './CalendarEntry';
 export * from './CalendarPanel';
-export * from './Day';
-export * from './DayDrawer';
 export * from './MultiFieldEditor';
